# swagger.yaml
host: "localhost:8080"
schemes:
    - http
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
        description: enter the token with the Bearer prefix e.g. Bearer eyJhbGci...
definitions:
...
    actor:
        title: actorResponse represents the response body for an actor.
        type: object
        properties:
            birthday:
                description: The birthday of the actor.
                example: "2000-01-01T00:00:00Z"
                format: date-time
                type: string
            gender:
                description: 'The gender of the actor, can be: ["male", "female", "other"].'
                example: male
                type: string
            id:
                description: The ID of the actor.
                example: 1
                format: int32
                type: integer
            name:
                description: The name of the actor.
                example: Leonardo Di Caprio
                type: string
    actorsWithMovies:
        type: array
        items:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                    example: 1
                name:
                    type: string
                    example: Leonardo Di Caprio
                    description: The name of the actor.
                gender:
                    type: string
                    example: male
                    description: The gender of the actor.
                birthday:
                    type: string
                    example: "2000-01-01T00:00:00Z"
                    description: The birthday of the actor.
                movies:
                    type: array
                    items:
                        $ref: '#/definitions/movie'

        title: allActorsResponse represents the response body for all actors.
    actorRequest:
        type: object
        title: actorRequest represents the request body for an actor.
        properties:
            name:
                description: The name of the actor.
                example: Leonardo Di Caprio
                type: string
            birthday:
                description: The birthday of the actor.
                example: "2000-01-01T00:00:00Z"
                format: date-time
                type: string
            gender:
                description: 'The gender of the actor, can be: ["male", "female", "other"].'
                example: male
                type: string
...
    movie:
        properties:
            description:
                description: The description of the movie.
                example: A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers.
                type: string
            id:
                description: The ID of the movie.
                example: 123
                format: int32
                type: integer
            name:
                description: The name of the movie.
                example: The Matrix
                type: string
            rating:
                description: The rating of the movie.
                example: "8.7"
                type: string
            release_date:
                description: The release date of the movie.
                example: "2012-07-16T00:00:00Z"
                format: date-time
                type: string
        type: object
        title: movieResponse represents the response for a movie.
...
info: {}
paths:
...
    /actors:
        delete:
            security:
                - Bearer: []
            consumes: 
                - application/json
            operationId: deleteActor
            parameters: 
                - in: query
                  name: id
                  required: true
                  description: The ID of the actor to be deleted.
            produces:
                - application/json
            responses:
                200:
                    schema:
                        type: object
                        properties:
                            id:
                                type: integer
                                format: int32
                                example: 1
                400:
                    $ref: '#/responses/error400Response'
                403:
                    $ref: '#/responses/error403Response'
                404:
                    $ref: '#/responses/error404Response'
            summary: Deletes an actor by ID, ONLY FOR ADMINISTRATORS.
            tags:
                - actors
        post:
            security:
                - Bearer: []
            operationId: createActor
            parameters:
                - in: body
                  name: actor
                  schema:
                      $ref: '#/definitions/actorRequest'
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/actor'
                "400":
                    $ref: '#/responses/error400Response'
                "401":
                    $ref: '#/responses/error401Response'
                "500":
                    $ref: '#/responses/error500Response'
            summary: Creates a new actor, ONLY FOR ADMINISTRATORS.
            tags:
                - actors
        patch:
            security:
                - Bearer: []
            operationId: updateActor
            consumes:
                - application/json
            parameters:
                - in: body
                  name: actor
                  schema:
                      $ref: '#/definitions/actor'
            produces:
                - application/json
            responses:
                200:
                    $ref: '#/responses/actor'
                400:
                    $ref: '#/responses/error400Response'
                403:
                    $ref: '#/responses/error403Response'
                404:
                    $ref: '#/responses/error404Response'
                500:
                    $ref: '#/responses/error500Response'
            summary: Updates an existing actor, ONLY FOR ADMINISTRATORS.
            tags:
                - actors
    /actors-movies:
        get:
            security:
                - Bearer: []
            operationId: actorsWithMovies
            produces:
                - application/json
            summary: Gets all actors with their movies.
            tags:
                - actors
            responses:
                200:
                    $ref: '#/responses/actorsWithMovies'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
...
responses:
    actor:
        description: actorResponse represents the response body for an actor.
        properties:
            actor:
                description: The actor information.
        schema:
            $ref: '#/definitions/actor'
    actorsWithMovies:
        description: actorsWithMoviesResponse represents the response body for all actors with movies.
        properties:
            actors:
                description: The list of actors.
        schema:
            $ref: '#/definitions/actorsWithMovies'    
...
    error400Response:
        description: Bad Request.
    error401Response:
        description: Unauthorized.
    error403Response:
        description: Forbidden.
    error404Response:
        description: Not Found.
    error500Response:
        description: Internal Server Error.
swagger: "2.0"